some small error fixes

+++ zeta2/zeta2.h
@@ -6,8 +6,8 @@ double zeta2 (int nthreads, int n) {
     double sum = 0;
 
     #pragma omp parallel for num_threads(nthreads) reduction(+:sum)
-    for (double i=1; i<=n; i++) {
-        sum += 1/(i*i);
+    for (int i=1; i<=n; i++) {
+        sum += 1.0/((double)i*(double)i);
     }
 
     double result = sqrt(sum*6);
